requirements.txt
alembic==1.14.1
annotated-types==0.7.0
asyncpg==0.30.0
bcrypt==4.3.0
black==25.1.0
fastapi==0.115.8
Jinja2==3.1.5
passlib==1.7.4
psycopg2-binary==2.9.10
pydantic[email]==2.10.6
pydantic-settings==2.8.0
PyJWT==2.10.1
SQLAlchemy==2.0.38
typer==0.15.1
uvicorn==0.34.0
websockets==15.0
